:truck: Rename: common 폴더로 통합
