+++ src/main/java/gaji/service/global/common/exception/code/BaseCodeDto.java
@@ -1,4 +1,4 @@
-package gaji.service.global.exception.code;
+package gaji.service.global.common.exception.code;
 
 import lombok.Builder;
 import lombok.Getter;
